requirements.txt
# Core AI
anthropic>=0.49.0

# Web Scraping
firecrawl-py>=1.4.0

# Scheduling
APScheduler>=3.10.0
pytz>=2024.1

# Config
python-dotenv>=1.0.0

# HTTP (Telegram delivery)
requests>=2.32.0

# Testing
pytest>=8.0.0
pytest-mock>=3.14.0
